feat: add subcommands
One for the tui and one for rendering a single file
